[WIP] concurrent lock-free queue + hazard pointer: first compilation cleaning (TODO: Use of hp for the queue then go for testing)

+++ libs/queue/queue.h
@@ -4,8 +4,7 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include <stdbool.h>
-
-#include "../memory_management/memory_management.h"
+#include "../SMR/hazard_pointer.h"
 
 struct queue_item{
    struct queue_item* next;
@@ -19,7 +18,7 @@ struct queue{
    size_t max_size;
 };
 
-struct queue* new_queue(uint32_t capacity);
+struct queue* new_queue(size_t capacity);
 
 bool queue_push(struct queue* queue, void* data);
